chore: flatten GraviteeConfigurationSecretResolver
