+++ gravitee-node-secrets/gravitee-node-secrets-service/src/main/java/io/gravitee/node/secrets/service/spring/SecretServiceConfiguration.java
@@ -3,7 +3,7 @@ package io.gravitee.node.secrets.service.spring;
 import io.gravitee.node.api.certificate.KeyStoreLoaderOptions;
 import io.gravitee.node.certificates.DefaultKeyStoreLoaderFactoryRegistry;
 import io.gravitee.node.secrets.plugins.SecretProviderPluginManager;
-import io.gravitee.node.secrets.service.conf.GraviteeConfigurationSecretResolverDispatcher;
+import io.gravitee.node.secrets.service.conf.GraviteeConfigurationSecretResolver;
 import io.gravitee.node.secrets.service.keystoreloader.SecretProviderKeyStoreLoaderFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -17,17 +17,17 @@ import org.springframework.core.env.Environment;
 public class SecretServiceConfiguration {
 
     @Bean
-    public GraviteeConfigurationSecretResolverDispatcher nodeSecretResolverDispatcher(
+    public GraviteeConfigurationSecretResolver nodeSecretResolverDispatcher(
         SecretProviderPluginManager secretProviderPluginManager,
         Environment environment
     ) {
-        return new GraviteeConfigurationSecretResolverDispatcher(secretProviderPluginManager, environment);
+        return new GraviteeConfigurationSecretResolver(secretProviderPluginManager, environment);
     }
 
     @Bean
     public SecretProviderKeyStoreLoaderFactory secretProviderKeyStoreLoaderFactory(
         DefaultKeyStoreLoaderFactoryRegistry<KeyStoreLoaderOptions> keyStoreLoaderFactoryRegistry,
-        GraviteeConfigurationSecretResolverDispatcher secretResolverDispatcher
+        GraviteeConfigurationSecretResolver secretResolverDispatcher
     ) {
         final SecretProviderKeyStoreLoaderFactory secretProviderKeyStoreLoaderFactory = new SecretProviderKeyStoreLoaderFactory(
             secretResolverDispatcher
