+++ gravitee-node-secrets/gravitee-node-secrets-service/src/test/java/io/gravitee/node/secrets/service/resolver/GraviteeConfigurationSecretPropertyResolverTest.java
@@ -4,7 +4,7 @@ import static io.gravitee.node.secrets.service.test.TestUtil.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.gravitee.node.secrets.plugins.internal.DefaultSecretProviderPluginManager;
-import io.gravitee.node.secrets.service.conf.GraviteeConfigurationSecretResolverDispatcher;
+import io.gravitee.node.secrets.service.conf.GraviteeConfigurationSecretResolver;
 import io.gravitee.secrets.api.core.Secret;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -28,7 +28,7 @@ class GraviteeConfigurationSecretPropertyResolverTest {
         DefaultSecretProviderPluginManager pluginManager = newPluginManager();
         MockEnvironment env = newEnvironment();
         env.setProperty("secrets.test.secrets.pass", "theMostPowerfulPasswordInTheWorldMate!!!");
-        GraviteeConfigurationSecretResolverDispatcher dispatcher = newDispatcher(pluginManager, env);
+        GraviteeConfigurationSecretResolver dispatcher = newDispatcher(pluginManager, env);
         this.cut = new GraviteeConfigurationSecretPropertyResolver(dispatcher);
     }
